backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-decouple==3.6
sqlparse==0.5.2
typing_extensions==4.12.2
djangorestframework==3.14.0